+++ src/components/qna/QuestionForm/index.ts
@@ -1,0 +1,1 @@
+export * from './QuestionForm'
