+++ src/components/qna/index.ts
@@ -1,3 +1,4 @@
+export * from './QaBadge'
 export * from './MarkdownEditor'
 export * from './AnswerCard'
 export * from './AnswerForm'
@@ -5,3 +6,6 @@ export * from './QuestionCard'
 export * from './CategoryFilter'
 export * from './CommentForm'
 export * from './CommentList'
+export * from './QuestionDetail'
+export * from './AnswerSection'
+export * from './QuestionForm'
